+++ sdk/lib/_internal/wasm/js_common/js_interop_patch.dart
@@ -729,6 +729,112 @@ extension StringToJSString on String {
   }
 }
 
+// -----------------------------------------------------------------------------
+// JSArray<JSNumber> <-> List<num>
+@patch
+extension JSArrayOfJSNumberToList on JSArray<JSNumber> {
+  @patch
+  List<double> get toDartDoubleList => [
+    for (var i = 0; i < this.length; i++) this[i].toDartDouble,
+  ];
+
+  @patch
+  List<int> get toDartIntList => [
+    for (var i = 0; i < this.length; i++) this[i].toDartInt,
+  ];
+}
+
+@patch
+extension ListOfNumberToJSArray on List<num> {
+  @patch
+  JSArray<JSNumber> get toJS => [for (var value in this) value.toJS].toJS;
+}
+
+// -----------------------------------------------------------------------------
+// JSArray<JSNumber?> <-> List<num?>
+@patch
+extension JSArrayOfNullableJSNumberToList on JSArray<JSNumber?> {
+  @patch
+  List<double?> get toDartDoubleList => [
+    for (var i = 0; i < this.length; i++) this[i]?.toDartDouble,
+  ];
+
+  @patch
+  List<int?> get toDartIntList => [
+    for (var i = 0; i < this.length; i++) this[i]?.toDartInt,
+  ];
+}
+
+@patch
+extension ListOfNullableNumberToJSArray on List<num?> {
+  @patch
+  JSArray<JSNumber?> get toJS => [for (var value in this) value?.toJS].toJS;
+}
+
+// -----------------------------------------------------------------------------
+// JSArray<JSString> <-> List<String>
+@patch
+extension JSArrayOfJSStringToList on JSArray<JSString> {
+  @patch
+  List<String> get toDartStringList => [
+    for (var i = 0; i < this.length; i++) this[i].toDart,
+  ];
+}
+
+@patch
+extension ListOfStringToJSArray on List<String> {
+  @patch
+  JSArray<JSString> get toJS => [for (var value in this) value.toJS].toJS;
+}
+
+// -----------------------------------------------------------------------------
+// JSArray<JSString?> <-> List<String?>
+@patch
+extension JSArrayOfNullableJSStringToList on JSArray<JSString?> {
+  @patch
+  List<String?> get toDartStringList => [
+    for (var i = 0; i < this.length; i++) this[i]?.toDart,
+  ];
+}
+
+@patch
+extension ListOfNullableStringToJSArray on List<String?> {
+  @patch
+  JSArray<JSString?> get toJS => [for (var value in this) value?.toJS].toJS;
+}
+
+// -----------------------------------------------------------------------------
+// JSArray<JSBoolean> <-> List<bool>
+@patch
+extension JSArrayOfJSBooleanToList on JSArray<JSBoolean> {
+  @patch
+  List<bool> get toDartBoolList => [
+    for (var i = 0; i < this.length; i++) this[i].toDart,
+  ];
+}
+
+@patch
+extension ListOfBoolToJSArray on List<bool> {
+  @patch
+  JSArray<JSBoolean> get toJS => [for (var value in this) value.toJS].toJS;
+}
+
+// -----------------------------------------------------------------------------
+// JSArray<JSBoolean?> <-> List<bool?>
+@patch
+extension JSArrayOfNullableJSBooleanToList on JSArray<JSBoolean?> {
+  @patch
+  List<bool?> get toDartBoolList => [
+    for (var i = 0; i < this.length; i++) this[i]?.toDart,
+  ];
+}
+
+@patch
+extension ListOfNullableBoolToJSArray on List<bool?> {
+  @patch
+  JSArray<JSBoolean?> get toJS => [for (var value in this) value?.toJS].toJS;
+}
+
 @patch
 extension JSAnyOperatorExtension on JSAny? {
   @patch