+++ tests/lib/js/static_interop_test/js_types_test.dart
@@ -4,6 +4,7 @@
 
 // Check that JS types work.
 
+import 'dart:collection';
 import 'dart:js_interop';
 import 'dart:js_interop_unsafe';
 import 'dart:typed_data';
@@ -56,6 +57,21 @@ external JSArray arr;
 
 @JS('arr')
 external JSArray<JSNumber> arrN;
+
+@JS()
+external JSArray<JSNumber?> arrNNullable;
+
+@JS()
+external JSArray<JSString> arrStr;
+
+@JS()
+external JSArray<JSString?> arrStrNullable;
+
+@JS()
+external JSArray<JSBoolean> arrBool;
+
+@JS()
+external JSArray<JSBoolean?> arrBoolNullable;
 
 @JS()
 external JSBoxedDartObject edo;
@@ -167,6 +183,32 @@ external JSAny? definedNonNullAny;
 
 @JS()
 external JSIterator<JSAny> getGenerator();
+
+class CustomList<E> extends ListBase<E> {
+  final List<E> _inner;
+
+  CustomList(this._inner);
+
+  int get length => _inner.length;
+
+  set length(int value) {
+    _inner.length = value;
+  }
+
+  E operator [](int index) => _inner[index];
+
+  void operator []=(int index, E value) {
+    _inner[index] = value;
+  }
+
+  void add(E value) {
+    _inner.add(value);
+  }
+
+  void addAll(Iterable<E> values) {
+    _inner.addAll(values);
+  }
+}
 
 class DartObject {
   String get foo => 'bar';
@@ -401,6 +443,233 @@ void syncTests() {
     Expect.notEquals(dartArrN, listN);
     Expect.equals(dartArrN.toJS, arrN);
   }
+
+  // [JSArray<JSNumber>] <-> [List<double>]
+  final listOfDoubles = <double>[1.0, 2.0];
+  final arrayOfDoubles = listOfDoubles.toJS;
+  Expect.isTrue(arrayOfDoubles is JSArray<JSNumber>);
+  Expect.isTrue(confuse(arrayOfDoubles) is JSArray<JSNumber>);
+
+  var dartArrayOfDoubles = arrayOfDoubles.toDartDoubleList;
+  Expect.equals(dartArrayOfDoubles.length, listOfDoubles.length);
+  Expect.equals(dartArrayOfDoubles[0], listOfDoubles[0]);
+  Expect.equals(dartArrayOfDoubles[1], listOfDoubles[1]);
+
+  if (isJSBackend) {
+    Expect.equals(dartArrayOfDoubles, listOfDoubles);
+    Expect.equals(dartArrayOfDoubles.toJS, listOfDoubles);
+  } else {
+    Expect.notEquals(dartArrayOfDoubles, listOfDoubles);
+    Expect.notEquals(dartArrayOfDoubles.toJS, listOfDoubles);
+  }
+
+  arrN = JSArray();
+  arrN.add((1.0).toJS);
+  dartArrayOfDoubles = arrN.toDartDoubleList;
+  Expect.equals(dartArrayOfDoubles.length, 1);
+  Expect.equals(dartArrayOfDoubles[0], 1.0);
+
+  Expect.throws(() => CustomList([1.0]).toJS);
+
+  // [JSArray<JSNumber>] <-> [List<int>]
+  final listOfInts = <int>[1, 2];
+  final arrayOfInts = listOfInts.toJS;
+  Expect.isTrue(arrayOfInts is JSArray<JSNumber>);
+  Expect.isTrue(confuse(arrayOfInts) is JSArray<JSNumber>);
+
+  final dartArrayOfInts = arrayOfInts.toDartIntList;
+  Expect.equals(dartArrayOfInts.length, listOfInts.length);
+  Expect.equals(dartArrayOfInts[0], listOfInts[0]);
+  Expect.equals(dartArrayOfInts[1], listOfInts[1]);
+
+  if (isJSBackend) {
+    Expect.equals(dartArrayOfInts, listOfInts);
+    Expect.equals(dartArrayOfInts.toJS, listOfInts);
+  } else {
+    Expect.notEquals(dartArrayOfInts, listOfInts);
+    Expect.notEquals(dartArrayOfInts.toJS, listOfInts);
+  }
+
+  Expect.throws(() => CustomList([1]).toJS);
+
+  // Copy the list to elicit an error even on platforms where this only throws
+  // lazily.
+  Expect.throws(() => List.of([1.5.toJS].toJS.toDartIntList));
+
+  // [JSArray<JSNumber?>] <-> [List<double?>]
+  final listOfNullableDoubles = <double?>[1.0, null];
+  final arrayOfNullableDoubles = listOfNullableDoubles.toJS;
+  Expect.isTrue(arrayOfNullableDoubles is JSArray<JSNumber?>);
+  Expect.isTrue(confuse(arrayOfNullableDoubles) is JSArray<JSNumber?>);
+
+  var dartArrayOfNullableDoubles = arrayOfNullableDoubles.toDartDoubleList;
+  Expect.equals(
+    dartArrayOfNullableDoubles.length,
+    listOfNullableDoubles.length,
+  );
+  Expect.equals(dartArrayOfNullableDoubles[0], listOfNullableDoubles[0]);
+  Expect.equals(dartArrayOfNullableDoubles[1], listOfNullableDoubles[1]);
+
+  if (isJSBackend) {
+    Expect.equals(dartArrayOfNullableDoubles, listOfNullableDoubles);
+    Expect.equals(dartArrayOfNullableDoubles.toJS, listOfNullableDoubles);
+  } else {
+    Expect.notEquals(dartArrayOfNullableDoubles, listOfNullableDoubles);
+    Expect.notEquals(dartArrayOfNullableDoubles.toJS, listOfNullableDoubles);
+  }
+
+  arrNNullable = JSArray();
+  arrNNullable.add((1.0).toJS);
+  arrNNullable.add(null);
+  dartArrayOfNullableDoubles = arrNNullable.toDartDoubleList;
+  Expect.equals(dartArrayOfNullableDoubles.length, 2);
+  Expect.equals(dartArrayOfNullableDoubles[0], 1.0);
+  Expect.equals(dartArrayOfNullableDoubles[1], null);
+
+  Expect.throws(() => CustomList(<double?>[1.0]).toJS);
+
+  // [JSArray<JSNumber?>] <-> [List<int?>]
+  final listOfNullableInts = <int?>[1, null];
+  final arrayOfNullableInts = listOfNullableInts.toJS;
+  Expect.isTrue(arrayOfNullableInts is JSArray<JSNumber?>);
+  Expect.isTrue(confuse(arrayOfNullableInts) is JSArray<JSNumber?>);
+
+  final dartArrayOfNullableInts = arrayOfNullableInts.toDartIntList;
+  Expect.equals(dartArrayOfNullableInts.length, listOfNullableInts.length);
+  Expect.equals(dartArrayOfNullableInts[0], listOfNullableInts[0]);
+  Expect.equals(dartArrayOfNullableInts[1], listOfNullableInts[1]);
+
+  if (isJSBackend) {
+    Expect.equals(dartArrayOfNullableInts, listOfNullableInts);
+    Expect.equals(dartArrayOfNullableInts.toJS, listOfNullableInts);
+  } else {
+    Expect.notEquals(dartArrayOfNullableInts, listOfNullableInts);
+    Expect.notEquals(dartArrayOfNullableInts.toJS, listOfNullableInts);
+  }
+
+  Expect.throws(() => CustomList(<int?>[1]).toJS);
+
+  // Copy the list to elicit an error even on platforms where this only throws
+  // lazily.
+  Expect.throws(() => List.of([1.5.toJS, null].toJS.toDartIntList));
+
+  // [JSArray<JSString>] <-> [List<String>]
+  final listOfStrings = ["foo", "bar"];
+  final arrayOfStrings = listOfStrings.toJS;
+  Expect.isTrue(arrayOfStrings is JSArray<JSString>);
+  Expect.isTrue(confuse(arrayOfStrings) is JSArray<JSString>);
+
+  var dartArrayOfStrings = arrayOfStrings.toDartStringList;
+  Expect.equals(dartArrayOfStrings.length, listOfStrings.length);
+  Expect.equals(dartArrayOfStrings[0], listOfStrings[0]);
+  Expect.equals(dartArrayOfStrings[1], listOfStrings[1]);
+
+  if (isJSBackend) {
+    Expect.equals(dartArrayOfStrings, listOfStrings);
+    Expect.equals(dartArrayOfStrings.toJS, listOfStrings);
+  } else {
+    Expect.notEquals(dartArrayOfStrings, listOfStrings);
+    Expect.notEquals(dartArrayOfStrings.toJS, listOfStrings);
+  }
+
+  arrStr = JSArray();
+  arrStr.add("a".toJS);
+  dartArrayOfStrings = arrStr.toDartStringList;
+  Expect.equals(dartArrayOfStrings.length, 1);
+  Expect.equals(dartArrayOfStrings[0], "a");
+
+  Expect.throws(() => CustomList(["a"]).toJS);
+
+  // [JSArray<JSString?>] <-> [List<String?>]
+  final listOfNullableStrings = ["foo", null];
+  final arrayOfNullableStrings = listOfNullableStrings.toJS;
+  Expect.isTrue(arrayOfNullableStrings is JSArray<JSString?>);
+  Expect.isTrue(confuse(arrayOfNullableStrings) is JSArray<JSString?>);
+
+  var dartArrayOfNullableStrings = arrayOfNullableStrings.toDartStringList;
+  Expect.equals(
+    dartArrayOfNullableStrings.length,
+    listOfNullableStrings.length,
+  );
+  Expect.equals(dartArrayOfNullableStrings[0], listOfNullableStrings[0]);
+  Expect.equals(dartArrayOfNullableStrings[1], listOfNullableStrings[1]);
+
+  if (isJSBackend) {
+    Expect.equals(dartArrayOfNullableStrings, listOfNullableStrings);
+    Expect.equals(dartArrayOfNullableStrings.toJS, listOfNullableStrings);
+  } else {
+    Expect.notEquals(dartArrayOfNullableStrings, listOfNullableStrings);
+    Expect.notEquals(dartArrayOfNullableStrings.toJS, listOfNullableStrings);
+  }
+
+  arrStrNullable = JSArray();
+  arrStrNullable.add("a".toJS);
+  arrStrNullable.add(null);
+  dartArrayOfNullableStrings = arrStrNullable.toDartStringList;
+  Expect.equals(dartArrayOfNullableStrings.length, 2);
+  Expect.equals(dartArrayOfNullableStrings[0], "a");
+  Expect.equals(dartArrayOfNullableStrings[1], null);
+
+  Expect.throws(() => CustomList(<String?>["a"]).toJS);
+
+  // [JSArray<JSBoolean>] <-> [List<bool>]
+  final listOfBools = [true, false];
+  final arrayOfBooleans = listOfBools.toJS;
+  Expect.isTrue(arrayOfBooleans is JSArray<JSBoolean>);
+  Expect.isTrue(confuse(arrayOfBooleans) is JSArray<JSBoolean>);
+
+  var dartArrayOfBooleans = arrayOfBooleans.toDartBoolList;
+  Expect.equals(dartArrayOfBooleans.length, listOfBools.length);
+  Expect.equals(dartArrayOfBooleans[0], listOfBools[0]);
+  Expect.equals(dartArrayOfBooleans[1], listOfBools[1]);
+
+  if (isJSBackend) {
+    Expect.equals(dartArrayOfBooleans, listOfBools);
+    Expect.equals(dartArrayOfBooleans.toJS, listOfBools);
+  } else {
+    Expect.notEquals(dartArrayOfBooleans, listOfBools);
+    Expect.notEquals(dartArrayOfBooleans.toJS, listOfBools);
+  }
+
+  arrBool = JSArray();
+  arrBool.add(true.toJS);
+  dartArrayOfBooleans = arrBool.toDartBoolList;
+  Expect.equals(dartArrayOfBooleans.length, 1);
+  Expect.equals(dartArrayOfBooleans[0], true);
+
+  Expect.throws(() => CustomList([true]).toJS);
+
+  // [JSArray<JSBoolean>] <-> [List<bool>]
+  final listOfNullableBools = [true, null];
+  final arrayOfNullableBooleans = listOfNullableBools.toJS;
+  Expect.isTrue(arrayOfNullableBooleans is JSArray<JSBoolean?>);
+  Expect.isTrue(confuse(arrayOfNullableBooleans) is JSArray<JSBoolean?>);
+
+  var dartArrayOfNullableBooleans = arrayOfNullableBooleans.toDartBoolList;
+  Expect.equals(
+    dartArrayOfNullableBooleans.length,
+    listOfNullableBools.length,
+  );
+  Expect.equals(dartArrayOfNullableBooleans[0], listOfNullableBools[0]);
+  Expect.equals(dartArrayOfNullableBooleans[1], listOfNullableBools[1]);
+
+  if (isJSBackend) {
+    Expect.equals(dartArrayOfNullableBooleans, listOfNullableBools);
+    Expect.equals(dartArrayOfNullableBooleans.toJS, listOfNullableBools);
+  } else {
+    Expect.notEquals(dartArrayOfNullableBooleans, listOfNullableBools);
+    Expect.notEquals(dartArrayOfNullableBooleans.toJS, listOfNullableBools);
+  }
+
+  arrBoolNullable = JSArray();
+  arrBoolNullable.add(true.toJS);
+  arrBoolNullable.add(null);
+  dartArrayOfNullableBooleans = arrBoolNullable.toDartBoolList;
+  Expect.equals(dartArrayOfNullableBooleans.length, 2);
+  Expect.equals(dartArrayOfNullableBooleans[0], true);
+  Expect.equals(dartArrayOfNullableBooleans[1], null);
+
+  Expect.throws(() => CustomList(<bool?>[true]).toJS);
 
   // [ArrayBuffer] <-> [ByteBuffer]
   buf = Uint8List.fromList([0, 255, 0, 255]).buffer.toJS;